Revert "clean up and added file structure"
