+++ package.json
@@ -17,7 +17,6 @@
     "crypto": "^1.0.1",
     "dotenv": "^16.1.3",
     "express": "^4.17.3",
-    "express-list-endpoints": "^6.0.0",
     "mongoose": "^7.2.1",
     "nodemon": "^2.0.15",
     "openai": "^3.2.1"
